Remove narwhal manager
